requirements.txt
discord
pillow
requests
imagetext_py